% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datademo.R
\docType{data}
\name{datademo}
\alias{datademo}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A data frame with 1170 rows and 13 variables:
\describe{
  \item{year}{year of 2016}
  \item{month}{June}
  \item{day}{1st to 30th}
  \item{hour}{0 to 23, in hour}
  \item{minute}{0 and 30, in minute}
  \item{H}{observed sensible heat, in W/m2}
  \item{LE}{observed latent heat, in W/m2}
  \item{TA}{air temperature, in deg C}
  \item{TS}{surface temperature, in deg C}
  \item{NETRAD}{net radiation, in W/m2}
  \item{RH}{Relative humidity, in percent}
  \item{G}{observed ground heat, in W/m2}
  \item{PA}{atmospheric pressure, in Kpa}
  ...
}
}
\source{
\url{http://www.datademo.info/}
}
\usage{
datademo
}
\description{
A dataset containing the time series with variables of almost 1170.
datademo.
}
\keyword{datasets}
