% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETannual.R
\docType{data}
\name{ETannual}
\alias{ETannual}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{ETannual}{annaul ET, in W/m2}
  ...
}
}
\usage{
ETannual
}
\description{
Global annual average ET during 1978-2018
ETannual.
}
\keyword{datasets}
