% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I.R
\docType{data}
\name{I}
\alias{I}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{I}{thermal inertia, in tiu}
  ...
}
}
\usage{
I
}
\description{
The surface inertia
I.
}
\keyword{datasets}
