% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat.R
\docType{data}
\name{lat}
\alias{lat}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A vector of latitude
\describe{
  \item{lat}{latitude, in deg}
  ...
}
}
\usage{
lat
}
\description{
Latitude
lat.
}
\keyword{datasets}
