% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lon.R
\docType{data}
\name{lon}
\alias{lon}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A vector of longitude
\describe{
  \item{lat}{longitude, in deg}
  ...
}
}
\usage{
lon
}
\description{
Longitude
lat.
}
\keyword{datasets}
