% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEP.R
\docType{data}
\name{MEP}
\alias{MEP}
\title{Variable for MEP model example.}
\format{
A vector of MEP estimations
\describe{
  \item{MEP}{MEP estimations, in W/m2}
  ...
}
}
\usage{
MEP
}
\description{
A vector of MEP estimations
MEP.
}
\keyword{datasets}
