% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{data}
\name{obs}
\alias{obs}
\title{Variable for MEP model example.}
\format{
A vector of observed latent heat flux
\describe{
  \item{obs}{observations, in W/m2}
  ...
}
}
\usage{
obs
}
\description{
A vector of observations
obs.
}
\keyword{datasets}
