% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\docType{data}
\name{qs}
\alias{qs}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{qs}{specific humidity, in kg/kg}
  ...
}
}
\usage{
qs
}
\description{
Specific humidity in 2000/01
qs.
}
\keyword{datasets}
