% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMEP-package.R
\docType{package}
\name{RMEP-package}
\alias{RMEP-package}
\title{RMEP: Calculating Evapotranspiration and Energy Fluxes by Maximum Entropy Production Model}
\description{
Accurate estimate of energy budget and evapotranspiration are significant in global ecosystem and climate.This package, based on the maximum entropy production(MEP) model, providing functions for calculating latent heat flux, sensible heat flux,ground heat flux as well as evapotranspiration(actual ET and potential ET), which can be applied in the fields of hydrology, meteorology and geophysics.
}
\keyword{internal}
