% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datapreprocessing.R
\name{RMEP_mice}
\alias{RMEP_mice}
\title{Multiple imputation methods for multivariate missing data}
\usage{
RMEP_mice(data, method = "pmm")
}
\arguments{
\item{data}{A dataframe of variables but contains missing data for MEP inputs}

\item{method}{Methods for imputation, including six methods selected}
}
\value{
A dataframe with no missing values
}
\description{
Deal with the missing data of the inputs for the maximum entropy production(MEP) model.
The function creates multiple imputations (replacement values) for multivariate missing data.
The imputation methods are established based on the mice package:Generates Multivariate Imputations by Chained Equations (MICE).
There are mainly six univariate imputation methods available including:Predictive mean matching("pmm"), Weighted predictive mean matching("midastouch"),
Classification and regression trees ("cart"), Unconditional mean imputation ("mean"), Imputation of quadratic terms("quadratic") and Linear regression, predicted values ("norm.predict").
Users can select method for imputations and the default is Predictive mean matching("pmm").
}
\examples{
RMEP_mice(airquality,method= 'pmm')
}
