% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMEP.R
\name{RMEP_nc}
\alias{RMEP_nc}
\title{Implementing MEP model with NetCDF format data}
\usage{
RMEP_nc(Rn, RnL, qs, Ts, I, z, type)
}
\arguments{
\item{Rn}{net radiation (unit:W/m2)}

\item{RnL}{net long-wave radiation (unit:W/m2)}

\item{qs}{specific humidity (unit:kg/kg)}

\item{Ts}{surface temperature (unit:Celsius)}

\item{I}{media thermal inertial (units:TIU)}

\item{z}{theoretical height above surface (unit:m)}

\item{type}{1 for bare soil surface or short canopy, 2 for dense canopy and 3 for Water-snow-ice surface}
}
\value{
Latent heat flux (unit:W/m2)
}
\description{
Calculate actual evapotranspiration by the MEP model with data of NetCDF format. The dimensions of all inputs must be consistent.
}
\examples{
RMEP_nc(Rn, RnL, qs, Ts, I, z, type)
}
