% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotfunction.R
\name{RMEP_plot_sca}
\alias{RMEP_plot_sca}
\title{Scatter plot between MEP estimates and observations}
\usage{
RMEP_plot_sca(obs, MEP, labx, laby)
}
\arguments{
\item{obs}{A vector of observations}

\item{MEP}{A vector of MEP estimations}

\item{labx}{Label for X-axis}

\item{laby}{Label for Y-axis}
}
\description{
Generating a scatter plot with MEP estimations and observations. The labels contains: label plot with the regression equation (a fitted polynomial model) and label with coefficient of determination (R^2) for fitted model.
}
\examples{
RMEP_plot_sca(MEP,obs)
}
