% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotfunction.R
\name{RMEP_plot_spa}
\alias{RMEP_plot_spa}
\title{Spatial distribution map of MEP estimated evapotranspiration}
\usage{
RMEP_plot_spa(
  lon,
  lat,
  MEP,
  main = "Latent heat flux by MEP",
  legends = "W m-2"
)
}
\arguments{
\item{lon}{The varaible of longitude}

\item{lat}{The variable of latitude}

\item{MEP}{The variable of MEP estimated evapotranspiration}

\item{main}{The title of figure}

\item{legends}{The legend of figure}
}
\description{
Generating a map for spatial distribution of evapotranspiration.The plot includes image and legend, generated based on the fields package.
}
\examples{
RMEP_plot_spa(lon,lat,ETannual)
}
