% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMEP.R
\name{RMEPPET}
\alias{RMEPPET}
\title{Calculate potential ET}
\usage{
RMEPPET(Rn, RnL, Ts, type)
}
\arguments{
\item{Rn}{net radiation(unit:W/m2)}

\item{RnL}{net long-wave radiation(unit:W/m2)}

\item{Ts}{surface temperature(unit:Celsius)}

\item{type}{1 for bare soil surface or short canopy, 2 for dense canopy and 3 for Water-snow-ice surface}
}
\value{
A list concludes latent heat flux(EMEP),sensible heat flux(HMEP),ground heat flux(GMEP) and potential ET(PETMEP)
}
\description{
Calculate energy fluxes and potential evapotranspiration based on the MEP model
}
\examples{
RMEPPET(Rn=200,RnL=100,Ts=20,type=1)
}
