% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rn.R
\docType{data}
\name{Rn}
\alias{Rn}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{Rn}{net radaition, in W/m2}
  ...
}
}
\usage{
Rn
}
\description{
Net radiation in 2000/01
Rn.
}
\keyword{datasets}
