% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnL.R
\docType{data}
\name{RnL}
\alias{RnL}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{RnL}{net Long-wave radaition, in W/m2}
  ...
}
}
\usage{
RnL
}
\description{
Net Long-wave radiation in 2000/01
RnL.
}
\keyword{datasets}
