% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shum.R
\name{Shum}
\alias{Shum}
\title{calculate specific humidity for MEP model input}
\usage{
Shum(TA, PA, RH)
}
\arguments{
\item{TA}{Air temperature(unit: deg C)}

\item{PA}{Atmospheric pressure(unit: kPa)}

\item{RH}{Relative humidity(unit: percent)}
}
\value{
specific humidity(unit:kg/kg)
}
\description{
calculate specific humidity for MEP model input
}
\examples{
Shum(TA=20,PA=101,RH=50)
}
