% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\docType{data}
\name{sm}
\alias{sm}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{sm}{soil miosture, in kg/m2}
  ...
}
}
\usage{
sm
}
\description{
Soil moisture in 2000/01
sm.
}
\keyword{datasets}
