% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Variable for MEP model example.}
\format{
A dataframe includes 3 variables
\describe{
  \item{time}{time series}
  \item{MEP}{MEP estimations, in W/m2}
  \item{obs}{observations, in W/m2}
  ...
}
}
\usage{
timeseries
}
\description{
A dataframe includes 3 variables
timeseries.
}
\keyword{datasets}
