% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ts.R
\docType{data}
\name{Ts}
\alias{Ts}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{Ts}{surface temperature, in deg C}
  ...
}
}
\usage{
Ts
}
\description{
Surface temperature in 2000/01
Ts.
}
\keyword{datasets}
