% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type.R
\docType{data}
\name{type}
\alias{type}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{type}{surface type, 1 or 2 or 3}
  ...
}
}
\usage{
type
}
\description{
Distance above the surface
type.
}
\keyword{datasets}
