% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{variables}
\alias{variables}
\title{Variables for MEP model analysis in Vignettes.}
\description{
Variables for MEP model analysis in Vignettes.
}
\details{
- `lat`: latitude, in deg
- `lon`: longitude, in deg
- `Ts` : surface temperature, in deg C
- `qs` : specific humidity, in kg/kg
- `sm`  : soil miosture, in kg/m2

- `I`  : surface thermal inertia, in tiu
- `Rn` : net radaition, in W/m2
- `RnL`: net Long-wave radaition, in W/m2
- `MEP`: MEP estimations, in W/m2
- `obs`: ET observations, in W/m2
}
