% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\docType{data}
\name{z}
\alias{z}
\title{Variables for MEP model analysis in Vignettes.}
\format{
A matrix in Netcdf format
\describe{
  \item{z}{Distance above the surface, in m}
  ...
}
}
\usage{
z
}
\description{
Distance above the surface
z.
}
\keyword{datasets}
