% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMEP.R
\name{RMEPPET}
\alias{RMEPPET}
\title{Calculate potential ET}
\usage{
RMEPPET(Rn, RnL, Ts, I = 800, z = 2.5, type = 1)
}
\arguments{
\item{Rn}{net radiation(unit:W/m2)}

\item{RnL}{net long-wave radiation(unit:W/m2)}

\item{Ts}{surface temperature(unit:Celsius)}

\item{I}{media thermal inertial (units:TIU),default is 600 TIU}

\item{z}{theoretical height above surface (unit:m),constant value and default is 2.5 m}

\item{type}{1 for bare soil surface or short canopy, 2 for dense canopy and 3 for Water-snow-ice surface}
}
\value{
A list concludes latent heat flux(EMEP,W/m2),sensible heat flux(HMEP,W/m2),ground heat flux(GMEP,W/m2) and potential ET(PETMEP,W/m2)
}
\description{
Calculate energy fluxes and potential evapotranspiration based on the MEP model
}
\examples{
RMEPPET(Rn=200,RnL=100,Ts=20,type=1)
RMEPPET(Rn=300,RnL=100,Ts=25,I=800,z=8,type=1)
}
